license html and css files

+++ license.sh
@@ -2,15 +2,17 @@
 function applyLicense {
 	f=$(basename $1)
 	ext="${f##*.}"
-	filesize=$(wc -c < $1 | tr -d ' ')
-	dd if="$1" bs=1 count=$(wc -c < licenses/LICENSE.$ext | tr -d ' ') 2>/dev/null \
-		| diff licenses/LICENSE.$ext - &>/dev/null
-	if [[ $? -ne 0 ]] && [[ $filesize -gt 0 ]]; then
-		echo $1
-		if [[ $2 != '--check' ]]; then
-			cp $1 licenses/temp$$
-			cat licenses/LICENSE.$ext licenses/temp$$ > $1
-			rm licenses/temp$$
+	if [[ -e licenses/LICENSE.$ext ]]; then
+		filesize=$(wc -c < $1 | tr -d ' ')
+		dd if="$1" bs=1 count=$(wc -c < licenses/LICENSE.$ext | tr -d ' ') 2>/dev/null \
+			| diff licenses/LICENSE.$ext - &>/dev/null
+		if [[ $? -ne 0 ]] && [[ $filesize -gt 0 ]]; then
+			echo $1
+			if [[ $2 != '--check' ]]; then
+				cp $1 licenses/temp$$
+				cat licenses/LICENSE.$ext licenses/temp$$ > $1
+				rm licenses/temp$$
+			fi
 		fi
 	fi
 }
@@ -19,5 +21,4 @@ find . -type f ! -path './frontend/lib/*' \
                ! -path './venv/*' \
                ! -path './.git/*' \
                ! -path './frontend/dist/mongoWebShell.js' \
-               \( -name '*.py' -o -name '*.js' \) \
 	| while read file; do applyLicense "$file" $1; done
